refactor: move plugin handler to a proper folder + retry adjustments
